+++ plonecli/cli.py
@@ -8,6 +8,7 @@ import os
 from pkg_resources import WorkingSet
 from plonecli.exceptions import NotInPackageError
 from plonecli.exceptions import NoSuchValue
+from plonecli.exceptions import NoSuchPath
 from plonecli.registry import template_registry as reg
 
 
@@ -216,8 +217,11 @@ def run_serve(context):
 
 @cli.command('test')
 @click.option('-a', '--all', 'all', is_flag=True)
+@click.option('-t', '--test', 'test', is_flag=True)
+@click.argument('path', type=click.Path(), required=False)
+@click.option('-s', '--package', 'package', is_flag=True)
 @click.pass_context
-def run_test(context, all):
+def run_test(context, all, test, package, path):
     """Run the tests in your package"""
     if context.obj.get('target_dir', None) is None:
         raise NotInPackageError(context.command.name)
@@ -226,6 +230,17 @@ def run_test(context, all):
     ]
     if all:
         params.append('--all')
+    if test and not path:
+        raise NoSuchPath(context.command.name, path)
+    elif test and path:
+        params.append('--test')
+        params.append(path)
+    if package and path:
+        params.append('--package')
+        params.append(path)
+    elif package and not path:
+        raise NoSuchPath(context.command.name, path)
+
     echo(
         '\nRUN: {0}'.format(' '.join(params)),
         fg='green',