+++ CHANGES.rst
@@ -4,6 +4,9 @@ History
 0.2.2 (unreleased)
 ------------------
 
+- Add -t and -s option to the test command
+  [kakshay21]
+
 - Fix #33 ConfigParser.NoOptionError
 [kakshay21]